+++ extensions/zen-browser/src/actions/index.ts
@@ -63,3 +63,20 @@ export async function openNewWindow() {
 
   await runAppleScript(script);
 }
+
+export async function openNewPrivateWindow() {
+  const script = `
+    tell application "Zen" to activate
+
+    tell application "System Events"
+      repeat 10 times
+        if frontmost of process "Zen" then exit repeat
+        delay 0.1
+      end repeat
+
+      keystroke "p" using {command down, shift down}
+    end tell
+  `;
+
+  await runAppleScript(script);
+}